# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AIPPMCS - Adaptive Intelligent Water Pump Protection, Monitoring and Control System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /faults:
    get:
      operationId: getFaults
      tags: [faults]
      summary: Get all faults
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: active
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Faults list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fault"
  /faults/{id}:
    get:
      operationId: getFault
      tags: [faults]
      summary: Get a specific fault
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Fault detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fault"
...
  /schedules:
    get:
      operationId: getSchedules
      tags: [schedules]
      summary: Get all schedules
      responses:
        "200":
          description: Schedules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a new schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    put:
      operationId: updateSchedule
      tags: [schedules]
      summary: Update a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /control/command:
    post:
      operationId: sendControlCommand
      tags: [control]
      summary: Send a control command to the pump controller
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ControlCommandInput"
      responses:
        "200":
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponse"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, info]
        - name: limit
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /firmware/update:
    post:
      operationId: triggerFirmwareUpdate
      tags: [firmware]
      summary: Trigger a firmware OTA update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirmwareUpdateInput"
      responses:
        "200":
          description: Update triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        mfaToken:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, email, role, isActive, createdAt]
      properties:
        id:
          type: number
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [administrator, technician, maintenance, viewer]
        isActive:
          type: boolean
        mfaEnabled:
          type: boolean
        lastLogin:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    Fault:
      type: object
      required: [id, type, severity, cause, recommendedAction, occurrences, confidence, trend, isActive, firstSeen, lastSeen]
      properties:
        id:
          type: number
        type:
          type: string
          enum: [dry_running, over_current, under_voltage, locked_rotor, power_failure, brownout, frequent_restart, long_runtime, communication_failure, controller_overtemperature]
        severity:
          type: string
          enum: [critical, high, medium, low]
        cause:
          type: string
        recommendedAction:
          type: string
        occurrences:
          type: number
        confidence:
          type: number
          description: 0-100
        trend:
          type: string
          enum: [increasing, stable, decreasing]
        isActive:
          type: boolean
        firstSeen:
          type: string
          format: date-time
        lastSeen:
          type: string
          format: date-time
        description:
          type: ["string", "null"]
...
    Schedule:
      type: object
      required: [id, name, type, startTime, endTime, isActive, priority, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        type:
          type: string
          enum: [daily, weekly, monthly, holiday, vacation, manual_override, temporary_override]
        startTime:
          type: string
          description: HH:MM
        endTime:
          type: string
          description: HH:MM
        days:
          type: array
          items:
            type: number
            description: 0=Sunday, 6=Saturday
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        isActive:
          type: boolean
        priority:
          type: number
        createdAt:
          type: string
          format: date-time
    ScheduleInput:
      type: object
      required: [name, type, startTime, endTime]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [daily, weekly, monthly, holiday, vacation, manual_override, temporary_override]
        startTime:
          type: string
        endTime:
          type: string
        days:
          type: array
          items:
            type: number
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        isActive:
          type: boolean
          default: true
        priority:
          type: number
          default: 1
    ScheduleUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [daily, weekly, monthly, holiday, vacation, manual_override, temporary_override]
        startTime:
          type: string
        endTime:
          type: string
        days:
          type: array
          items:
            type: number
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        isActive:
          type: boolean
        priority:
          type: number
    ControlCommandInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [start, stop, restart, reset_fault, emergency_stop, enable_auto, enable_manual, disable_scheduling, lock_controls, unlock_controls]
        reason:
          type: ["string", "null"]
          description: Optional reason for logging
    CommandResponse:
      type: object
      required: [success, commandId, status]
      properties:
        success:
          type: boolean
        commandId:
          type: string
        status:
          type: string
          enum: [accepted, rejected, pending]
        message:
          type: ["string", "null"]
...
    Notification:
      type: object
      required: [id, type, severity, message, timestamp, isRead]
      properties:
        id:
          type: number
        type:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, info]
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        isRead:
          type: boolean
        details:
          type: ["string", "null"]
...
    FirmwareUpdateInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [update, rollback, enter_recovery]
        targetVersion:
          type: ["string", "null"]
...